requirements.txt
cloudinary==1.16.0
dj-database-url==0.5.0
Django==2.2.3

django-bootstrap-form==3.4
django-bootstrap-pagination==1.7.1
django-ckeditor==5.9.0
django-cloudinary-storage==0.2.3

django-taggit==1.2.0
gunicorn==19.9.0
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.1
    

